Adding integration tests for check in
